+++ include/cxx.h
@@ -1,4 +1,5 @@
 /*
+ * Copyright 2013 Esrille Inc.
  * Copyright 2008-2010 Google Inc.
  * Copyright 2007 Nintendo Co., Ltd.
  *
@@ -35,16 +36,16 @@ protected:
     bool useExceptions;
 
     bool constructorMode;
-    int optionalStage;
-    int optionalCount;
+    size_t optionalStage;
+    size_t optionalCount;
 
     std::string moduleName;
     const Node* currentNode;
 
-    int paramCount;  // The number of parameters of the previously evaluated operation
+    size_t paramCount;  // The number of parameters of the previously evaluated operation
     const ParamDcl* variadicParam;  // Non-NULL if the last parameter of the previously evaluated operation is variadic
 
-    int getParamCount() const
+    size_t getParamCount() const
     {
         return paramCount;
     }
@@ -580,7 +581,7 @@ public:
         {
             included = "No" + included;
         }
-        for (int i = 0; i < included.size(); ++i)
+        for (size_t i = 0; i < included.size(); ++i)
         {
             char c = included[i];
             included[i] = isalnum(c) ? (capitalize ? toupper(c) : c) : '_';
